feat: rename commands to be more user friendly
1. remove "stack describe"
2. rename "stack new" to "new"
3. rename "deployment apply" to "stack up"
4. rename "deployment delete" to "stack down"
5. rename "deployment list" to "stack list"
6. remove user aliases
