+++ pkg/cmd/root.go
@@ -20,15 +20,12 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
-	"github.com/mitchellh/mapstructure"
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/nitrictech/cli/pkg/cmd/deployment"
 	"github.com/nitrictech/cli/pkg/cmd/provider"
 	"github.com/nitrictech/cli/pkg/cmd/run"
 	cmdstack "github.com/nitrictech/cli/pkg/cmd/stack"
@@ -96,14 +93,14 @@ func init() {
 	})
 	cobra.CheckErr(err)
 
-	rootCmd.AddCommand(deployment.RootCommand())
-	rootCmd.AddCommand(provider.RootCommand())
+	newProjectCmd.Flags().BoolVarP(&force, "force", "f", false, "force project creation, even in non-empty directories.")
+	rootCmd.AddCommand(newProjectCmd)
 	rootCmd.AddCommand(cmdstack.RootCommand())
+	rootCmd.AddCommand(provider.RootCommand())
 	rootCmd.AddCommand(cmdTarget.RootCommand())
 	rootCmd.AddCommand(run.RootCommand())
 	rootCmd.AddCommand(versionCmd)
 	rootCmd.AddCommand(configHelpTopic)
-	addAliases()
 }
 
 // initConfig reads in config file and ENV variables if set.
@@ -131,12 +128,6 @@ func initConfig() {
 
 func ensureConfigDefaults() {
 	needsWrite := false
-	aliases := viper.GetStringMap("aliases")
-	if _, ok := aliases["new"]; !ok {
-		needsWrite = true
-		aliases["new"] = "stack create"
-		viper.Set("aliases", aliases)
-	}
 
 	to := viper.GetDuration("build_timeout")
 	if to == 0 {
@@ -165,26 +156,5 @@ func ensureConfigDefaults() {
 				return viper.WriteConfigAs(filepath.Join(utils.NitricConfigDir(), ".nitric-config.yaml"))
 			},
 			StopMsg: "Configfile updated"}, tasklet.Opts{})
-	}
-}
-
-func addAliases() {
-	aliases := map[string]string{}
-	cobra.CheckErr(mapstructure.Decode(viper.GetStringMap("aliases"), &aliases))
-	for n, aliasString := range aliases {
-		alias := &cobra.Command{
-			Use:   n,
-			Short: "alias for: " + aliasString,
-			Long:  "Custom alias command for " + aliasString,
-			Run: func(cmd *cobra.Command, args []string) {
-				newArgs := []string{os.Args[0]}
-				newArgs = append(newArgs, strings.Split(aliasString, " ")...)
-				newArgs = append(newArgs, args...)
-				os.Args = newArgs
-				cobra.CheckErr(rootCmd.Execute())
-			},
-			DisableFlagParsing: true, // the real command will parse the flags
-		}
-		rootCmd.AddCommand(alias)
 	}
 }
